# shared/contracts/core-api.openapi.yaml
openapi: 3.1.0
info: {title: ENQIVRA Core API, version: 0.2.0}
paths:
  /api/v1/auth/register: {post: {summary: Register user and owner organization, responses: {'201': {description: Created}}}}
  /api/v1/auth/login: {post: {summary: Exchange credentials for access and refresh tokens, responses: {'200': {description: Authenticated}}}}
  /api/v1/auth/refresh: {post: {summary: Rotate a refresh token, responses: {'200': {description: Rotated}}}}
  /api/v1/me: {get: {summary: Current user, security: [{bearerAuth: []}], responses: {'200': {description: User}}}}
  /api/v1/organizations: {get: {summary: Current memberships, security: [{bearerAuth: []}], responses: {'200': {description: Organizations}}}}
  /api/v1/organizations/{organizationId}/members: {get: {summary: List organization members, security: [{bearerAuth: []}], responses: {'200': {description: Members}}}, post: {summary: Add an existing registered user with a role, security: [{bearerAuth: []}], responses: {'201': {description: Member created}}}}
  /api/v1/assets: {get: {summary: List tenant assets, security: [{bearerAuth: []}], responses: {'200': {description: Assets}}}, post: {summary: Create asset, security: [{bearerAuth: []}], responses: {'201': {description: Created}}}}
  /api/v1/manufacturers: {get: {summary: List manufacturers, security: [{bearerAuth: []}], responses: {'200': {description: Manufacturers}}}, post: {summary: Create manufacturer, security: [{bearerAuth: []}], responses: {'201': {description: Created}}}}
  /api/v1/assets/{assetId}/components: {get: {summary: List components, security: [{bearerAuth: []}], responses: {'200': {description: Components}}}, post: {summary: Add component, security: [{bearerAuth: []}], responses: {'201': {description: Created}}}}
  /api/v1/cases: {get: {summary: List cases, security: [{bearerAuth: []}], responses: {'200': {description: Cases}}}, post: {summary: Create case, security: [{bearerAuth: []}], responses: {'201': {description: Created}}}}
  /api/v1/cases/{caseId}/evidence: {get: {summary: List evidence, security: [{bearerAuth: []}], responses: {'200': {description: Evidence}}}, post: {summary: Upload evidence, security: [{bearerAuth: []}], responses: {'201': {description: Created}}}}
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}